Organized Edit Roi module (ported from develop-feature)
