+++ studio/app/optinist/core/edit_ROI/__init__.py
@@ -1,3 +1,4 @@
-from studio.app.optinist.core.edit_ROI.edit_ROI import EditROI, EditRoiUtils
+from studio.app.optinist.core.edit_ROI.edit_ROI import EditROI
+from studio.app.optinist.core.edit_ROI.utils import EditRoiUtils
 
 __all__ = ["EditROI", "EditRoiUtils"]
